More legalisation work

+++ src/mir/ty.rs
@@ -1,7 +1,7 @@
 #[derive(Clone,Debug,PartialEq,Eq)]
 pub enum Type
 {
-    Integer { bit_width: u16 },
+    Integer { bit_width: u32 },
     Nothing,
 }
